Extract mock repository classes and do refactoring
