+++ lib/repository/pins_repository_mock.dart
@@ -1,0 +1,18 @@
+import 'package:mobile/model/pin_model.dart';
+import 'package:mobile/repository/pins_repository.dart';
+
+class MockPinsRepository extends PinsRepository {
+  factory MockPinsRepository() {
+    return _instance ?? MockPinsRepository._internal();
+  }
+
+  MockPinsRepository._internal();
+
+  static MockPinsRepository _instance;
+
+  @override
+  Future<List<Pin>> getHomePagePins({int page}) async {
+    final pins = List.filled(10, 1).map((_) => Pin.fromMock()).toList();
+    return Future.value(pins);
+  }
+}
