+++ lib/repository/repositories.dart
@@ -1,3 +1,4 @@
+export 'account_repository.dart';
 export 'boards_repository.dart';
 export 'pins_repository.dart';
 export 'users_repository.dart';
