+++ test/view/onboarding/authentication_bloc_test.dart
@@ -1,6 +1,6 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobile/data/account_repository.dart';
+import 'package:mobile/repository/account_repository.dart';
 import 'package:mobile/model/auth.dart';
 import 'package:mobile/view/onboarding/authentication_bloc.dart';
 import 'package:mockito/mockito.dart';
